+++ zucchini-ui-backend/src/main/java/io/zucchiniui/backend/scenario/domain/ScenarioMergeResult.java
@@ -1,0 +1,4 @@
+package io.zucchiniui.backend.scenario.domain;
+
+public class ScenarioMergeResult {
+}