+++ zucchini-ui-capsule/src/main/java/io/zucchiniui/capsule/ZucchiniUIApplication.java
@@ -7,9 +7,15 @@ import io.dropwizard.setup.Environment;
 import io.zucchiniui.backend.BackendBundle;
 import io.zucchiniui.backend.BackendConfiguration;
 import io.zucchiniui.backend.support.exceptionhandler.ExitExceptionHandler;
+import org.eclipse.jetty.servlet.FilterHolder;
 import org.eclipse.jetty.servlet.ServletHolder;
 
+import javax.servlet.DispatcherType;
+import java.util.EnumSet;
+
 public class ZucchiniUIApplication extends Application<BackendConfiguration> {
+
+    private static final String BASE_PATH = "/ui";
 
     public static void main(final String... args) throws Exception {
         final ZucchiniUIApplication application = new ZucchiniUIApplication();
@@ -20,19 +26,26 @@ public class ZucchiniUIApplication extends Application<BackendConfiguration> {
     @Override
     public void initialize(final Bootstrap<BackendConfiguration> bootstrap) {
         bootstrap.addBundle(new BackendBundle());
-        bootstrap.addBundle(new AssetsBundle("/ui", "/ui", "index.html", "ui"));
+        bootstrap.addBundle(new AssetsBundle("/ui", BASE_PATH, "index.html", "ui"));
     }
 
     @Override
     public void run(final BackendConfiguration configuration, final Environment environment) throws Exception {
         // Register the servlet that generates the UI Javascript config file
-        final ServletHolder uiConfigServletHolder = new ServletHolder(new UIConfigServlet(environment.getObjectMapper()));
-        environment.getApplicationContext().addServlet(uiConfigServletHolder, "/ui/config.js");
+        final ServletHolder uiConfigServletHolder = new ServletHolder(new UIConfigServlet(environment.getObjectMapper(), BASE_PATH));
+        environment.getApplicationContext().addServlet(uiConfigServletHolder, BASE_PATH + "/assets/config.js");
 
         // Redirect to UI
-        final ServletHolder redirectServletHolder = new ServletHolder(new RedirectServlet("/ui/"));
+        final ServletHolder redirectServletHolder = new ServletHolder(new RedirectServlet(BASE_PATH + "/"));
         environment.getApplicationContext().addServlet(redirectServletHolder, "/");
-        environment.getApplicationContext().addServlet(redirectServletHolder, "/ui");
+        environment.getApplicationContext().addServlet(redirectServletHolder, BASE_PATH);
+
+        // Forward 404 pages to index (used for browser history)
+        final FilterHolder forwardFilterHolder = new FilterHolder(new ForwardToIndexFilter(BASE_PATH));
+        environment.getApplicationContext().addFilter(forwardFilterHolder, BASE_PATH + "/*", EnumSet.allOf(DispatcherType.class));
+
+        // final ServletHolder forwardToIndexServletHolder = new ServletHolder(new ForwardToIndexServlet(BASE_PATH));
+        // environment.getApplicationContext().addServlet(forwardToIndexServletHolder, BASE_PATH + "/test-runs/*");
     }
 
 }