+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Zucchini UI Changelog
 Version 0.0.2-SNAPSHOT
 ----------------------
 
+* Grouping scenarios by error messages through regex(contribution by @ jeremiemarc)
 * Search in scenario content
 * Import only passed scenarios (contribution by @pierrick-boule)
 * Use Yarn instead of NPM for Javascript packages management