+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Zucchini UI Changelog
 =====================
 
+Version 0.0.3-SNAPSHOT
+----------------------
+
+* **New UI with React**
+* Search within a test run
+
+
 Version 0.0.2-SNAPSHOT
 ----------------------